Add daily_bounds to constraints.py
